VirusTotal API + Scapy testing
I started some python scripts to collect information from the VirusTotal API given a domain name. This can later be used once a DNS request is intercepted to see if the domain/URL is malicious according to VirusTotal if they are not on a block-list already.

I also did some testing with scapy in Python. This was used to collect DNS requests from the source device and write unique URLs to a file. This can also later be used to make dynamic block-lists.